feat: member kakaoId 타입 INTGER -> LONG 으로 변경 (#60)

+++ src/main/java/com/jungle/chalnaServer/domain/member/domain/entity/Member.java
@@ -33,7 +33,7 @@ public class Member extends BaseTimestampEntity {
     private String profileImageUrl = "/images/default_image.png";;
 
     @Column(nullable = false)
-    private Integer kakaoId;
+    private Long kakaoId;
 
     private String deviceId;
